# api/proto/url/url.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "url/url.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "URL"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/get/{Value}": {
      "get": {
        "operationId": "URL_ToFullLink",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiFullURL"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "Value",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "URL"
        ]
      }
    },
    "/post": {
      "post": {
        "operationId": "URL_ToShortLink",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiShortURL"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiFullURL"
            }
          }
        ],
        "tags": [
          "URL"
        ]
      }
    }
  },
  "definitions": {
    "apiFullURL": {
      "type": "object",
      "properties": {
        "Value": {
          "type": "string"
        }
      }
    },
    "apiShortURL": {
      "type": "object",
      "properties": {
        "Value": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}